Restructure TeacherView code in several files and fix flow errors
